+++ props/text_decoration_style.go
@@ -1,0 +1,15 @@
+package props
+
+type TextDecorationStyle string
+
+const (
+	TextDecorationStyleSolid  TextDecorationStyle = "solid"
+	TextDecorationStyleDouble TextDecorationStyle = "double"
+	TextDecorationStyleDotted TextDecorationStyle = "dotted"
+	TextDecorationStyleDashed TextDecorationStyle = "dashed"
+	TextDecorationStyleWavy   TextDecorationStyle = "wavy"
+)
+
+func (t TextDecorationStyle) String() string {
+	return string(t)
+}